+++ packages/webview_flutter/webview_flutter/lib/platform_interface.dart
@@ -22,5 +22,6 @@ export 'package:webview_flutter_platform_interface/webview_flutter_platform_inte
         WebSettings,
         WebResourceError,
         WebResourceErrorType,
+        WebViewCookie,
         WebViewRequest,
         WebViewRequestMethod;