+++ packages/webview_flutter/webview_flutter/CHANGELOG.md
@@ -1,6 +1,11 @@
-## 2.7.0
+## 2.8.0
 
 * Adds support for the `loadFlutterAsset` method.
+
+## 2.7.0
+
+* Adds `setCookie` to CookieManager.
+* CreationParams now supports setting `initialCookies`.
 
 ## 2.6.0